split files up
